Animation

In Springies.jl, you are limited only by your imagination. Since the resulting simulations differ widely in form, Springies.jl does not provide its own plotting routines, and it's up to you to decide how to do this. When simulations are large, one option is to save individual frames to disk and then assemble them into an animation using ffmpeg. For convenience, the two functions below are provided to minimise the need for boilerplate code when doing this:

Springies.make_framedir — Function

make_framedir(save_animation_as::String)

Create a directory to save animation frames.

Springies.run_ffmpeg — Function

run_ffmpeg(framedir, fps, save_as; naming="frame_%06d.png")

Run ffmpeg to assemble frames into a gif animation.

ffmpeg is run twice: pass 1 = generate palette, pass 2 = encode with palette