External forcing

Every ForceField type must have a corresponding applied_force function method which describes how the value of applied force at given coordinates. These are the aplied_force methods currently defined in Springies.jl:

Springies.applied_force — Function

applied_force(::ZeroForce{FT}, theta, dtheta, t)

Return zero.

applied_force(f::CosineForce{FT}, theta, dtheta, t)

Return value of basic cosine wave at time t.

applied_force(f::ClockForce{FT}, theta, dtheta, t)

Return F0 when pendulum is moving outwards and outside threshold θc, zero otherwise.

applied_force(f::DoubleGyre{FT}, x, y, t)

Return components of Double Gyre field in x and y directions.