Initial Conditions

Springies.jl defines some convenience functions for creating initial conditions:

Springies.meshgrid_xy — Function

meshgrid_xy(x, y)

Create grids from coordinates x and y.

x progresses to the right along the second dimension, and y progresses upwards along the first dimension.

Springies.init_particles — Function

init_particles(nx, ny, cx, cy, Ax, Ay; m:InitialisationMethod)

Initialise particle positions either as a regular grid or random placement.

Arguments

nx: number of particles along x dimension

ny: number of particles along y dimension

cx: x-coordinate centre of particle cluster

cy: y-coordinate centre of particle cluster

Ax: extent of particle cluster along x dimension

Ay: extent of particle cluster along y dimension

m: Grid() for grid, Random() for random placement

init_particles(n, cx, cy, Ax, Ay m::Random)

Convenience method for placing random particles, providing total number of particles n rather than particles in x and y directions.