System setups

This page shows how dynamical systems are encoded in Springies.jl, and shows the existing setup options provided.

Overview

Dynamical systems are encoded using Springy types, which contain the system parameters and external forcing. Each Springy type must have a corresponding differentials! function, which describes how the differential terms in the system are computed.

External forcing is given to Springy instances using ForceField types, for which a corresponding applied_force function must be defined.

Pendulums

We encode a pendulum as:

Springies.Pendulum1D — Type

Pendulum1D{FT,FF<:ForceField{FT}} <: Springy{FT}
Pendulum1D(m, c, L; g=9.81, F=ZeroForce)

A damped, driven pendulum in 1D. The keyword constructor defaults to g = 9.81 and zero external forcing.

Fields

m: Mass

c: Damping coefficient

L: Pendulum length

g: Gravitational acceleration

F: External force field

mL: Precomputed m * L

c_over_m: Precomputed c / m

g_over_L: Precomputed g / L

The following external forcings are available for use on a pendulum:

Springies.ZeroForce — Type

ZeroForce{FT} <: ForceField{FT}

Type for absence of external force.

For use with Pendulum1D springies.

Springies.CosineForce — Type

CosineForce{FT} <: ForceField{FT}

Type for a time-periodic external force of form F0 * cos(ω * t).

For use with Pendulum1D springies.

Springies.ClockForce — Type

ClockForce{FT} <: ForceField{FT}

Type for a crude clock-style forcing.

For use with Pendulum1D springies.

Fields

F0: Force amplitude

θc: Position threshold

The Double Gyre

Double Gyre problems use the Double Gyre field:

Springies.DoubleGyre — Type

DoubleGyre{FT} <: ForceField{FT}

The canonical Double Gyre field.

Fields

A: Velocity magnitude control

e: Wobble size control

omega: Wobble angular frequency

Two setups are currently available for use in this field:

Springies.FreeParticle2D — Type

FreeParticle2D{FT,FF<:ForceField{FT}} <: Springy{FT}

A massless particle that can be advected in two dimensions by a field.

Fields

F: External force field (conceptually this is a velocity field rather than a force field).

Springies.BendyStalk — Type

BendyStalk{FT,FF<:ForceField{FT}} <: Springy{FT}

A linear lumped mass model of a flexible mass-spring system. This is equivalent to two independent springs in the x and y directions.

Fields

m: Mass

c: Damping coefficient

k: Stiffness coefficent

x0: Neutral position in x

y0: Neutral position in y

F: External force field

c_over_m: Precomputed c / m

k_over_m: Precomputed k / m

The Three-body Problem

We encode the Three-body problem using this type:

Springies.ThreeBody — Type

ThreeBody{FT} <: Springy{FT}

The canonical Three-body Problem.

Fields

m1: Mass of particle 1

m2: Mass of particle 2

m3: Mass of particle 3

G: Gravitational constant