Types

Springies.jl defines two abstract supertypes; ForceField and Springy.

ForceField

ForceField structs contain the external forcing of a dynamical system. They are passed to Springy structs.

Springies.ForceField — Type

ForceField{FT<:AbstractFloat}

Abstract supertype for all external force fields.

Subtypes must implement callable functor giving value of force at input coordinates.

These are the concrete subtypes of ForceField currently defined in Springies.jl:

Springies.ZeroForce — Type

ZeroForce{FT} <: ForceField{FT}
(f::ZeroForce{FT})(θ, dθ, t)

Absence of external force. Return zero.

For use with Pendulum1D springies.

Springies.CosineForce — Type

CosineForce{FT} <: ForceField{FT}
(f::CosineForce)(θ, dθ, t)

A time-periodic external force. Return F0 * cos(ω * t).

For use with Pendulum1D springies.

Fields

F0: Force amplitude

omega: Angular frequency ω

Springies.ClockForce — Type

ClockForce{FT} <: ForceField{FT}
(f::ClockForce)(θ, dθ, t)

An external force that is applied only when pendulum is moving outwards and outside a threshold θc. Return F0 when these criteria are met, and zero otherwise.

For use with Pendulum1D springies.

Fields

F0: Force amplitude

θc: Position threshold

Springies.DoubleGyre — Type

DoubleGyre{FT} <: ForceField{FT}
(f::DoubleGyre)(x, y, t)

The canonical Double Gyre field. Functor returns [u, v], the x and y components of the field, respectively.

Fields

A: Velocity magnitude control

e: Wobble size control

omega: Wobble angular frequency

Springy

Springy structs contain the parameters and external forcing of a dynamical system. The external forcing is contained within a ForceField struct.

Springies.Springy — Type

Springy{FT<:AbstractFloat}

Abstract supertype for all Springies.

Subtypes must implement differentials!(du, u, p::MyNewSpringy{FT}, t).

These are the concrete subtypes of Springy currently defined in Springies.jl:

Springies.Pendulum1D — Type

Pendulum1D{FT} <: Springy{FT}
Pendulum1D(m, c, L; g=9.81, F=ZeroForce)

A damped, driven pendulum in 1D. The keyword constructor defaults to g = 9.81 and zero external forcing.

Fields

m: Mass

c: Damping coefficient

L: Pendulum length

g: Gravitational acceleration

F: External force field

mL: Precomputed m * L

c_over_m: Precomputed c / m

g_over_L: Precomputed g / L

Springies.FreeParticle2D — Type

FreeParticle2D{FT} <: Springy{FT}

A massless particle that can be advected in two dimensions by a field.

Fields

F: External force field (conceptually this is a velocity field rather than a force field).

Springies.BendyStalk — Type

BendyStalk{FT} <: Springy{FT}

A linear lumped mass model of a flexible mass-spring system. This is equivalent to two independent springs in the x and y directions.

Fields

m: Mass

c: Damping coefficient

k: Stiffness coefficent

x0: Neutral position in x

y0: Neutral position in y

F: External force field

c_over_m: Precomputed c / m

k_over_m: Precomputed k / m

Springies.ThreeBody — Type

ThreeBody{FT} <: Springy{FT}

The canonical Three-body Problem.

Fields

m1: Mass of particle 1

m2: Mass of particle 2

m3: Mass of particle 3

G: Gravitational constant